pyproject.toml
[tool.poetry]
name = "pythonapi"
version = "0.1.0"
description = "external API for telegram bot"
authors = ["Alena Meshcheriakova"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
fastapi-cache = "^0.1.0"
fastapi-users = "^13.0.0"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.4.0"
pytest = "^8.3.2"
pytest-dotenv = "^0.5.2"
pytest-mock = "^3.14.0"
grpcio = "^1.65.4"
protobuf = "^5.27.3"
httpx = "^0.27.0"
pytest-grpc = "^0.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
